Change how NaNs are imputed
We should use the mode (i.e., most common value) of the column for
categorical variables, and the median for continuous variables.

Since there’s no easy way to detect continuous vs. categorical
variables in pandas, we use a heuristic: If >20% of the values in a
column are unique, then it is probably a continuous variable.
Otherwise, it is probably a categorical variable.

(Related to #1)